Texture abstraction
G rendu possible de pouvoir mettre des texture custom depuis l'app
